// src/app/auth/sign-up/sign-up-form-petitioners/models/sign-up-petitioners-request.model.ts
export interface SignUpPetitionersRequest {
  cuit: string;
  email: string;
  password: string;
  passwordRepeat: string;
  businessName: string;
  phone: string;
  street: string;
  streetNumber: string;
  apartment: string;
  postCode: string;
  localityId: number;
  provinceId: number;
  name: string;
  lastName: string;
  emailAccountManager: string;
  phoneAccountManager: string;
  termsAndConditions: string;
  tokenCaptcha: string;
}